feat: Add cmake support for webrtc gloox

+++ src/lib/messenger/IMFile.h
@@ -4,21 +4,21 @@
  * You can use this software according to the terms and conditions of the Mulan
  * PubL v2. You may obtain a copy of Mulan PubL v2 at:
  *          http://license.coscl.org.cn/MulanPubL-2.0
- * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
- * EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
- * MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
- * See the Mulan PubL v2 for more details.
+ * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY
+ * KIND, EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
+ * NON-INFRINGEMENT, MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE. See the
+ * Mulan PubL v2 for more details.
  */
 #ifndef IMFILE_H
 #define IMFILE_H
 
 #include "IM.h"
+#include "lib/messenger/messenger.h"
 #include <QFile>
 #include <QThread>
-#include <gloox/bytestreamdatahandler.h>
-#include <gloox/inbandbytestream.h>
-#include <gloox/iqhandler.h>
-#include "lib/messenger/messenger.h"
+#include <gloox/src/bytestreamdatahandler.h>
+#include <gloox/src/inbandbytestream.h>
+#include <gloox/src/iqhandler.h>
 
 namespace lib {
 namespace messenger {
@@ -26,8 +26,8 @@ namespace messenger {
 class IMFile : public QThread, public BytestreamDataHandler {
   Q_OBJECT
 public:
-  IMFile(const JID &friendId,    //
-        FileHandler::File file, //
+  IMFile(const JID &friendId,           //
+         FileHandler::File file,        //
          const lib::messenger::IM *im); //
 
   virtual ~IMFile();
@@ -46,8 +46,8 @@ public:
 
   void handleBytestreamClose(Bytestream *bs) override;
 
-  bool sendFinished()const {return m_file.size == m_sentBytes;}
-  bool ackFinished()const {return m_seq > 0 && m_seq == m_ack_seq;}
+  bool sendFinished() const { return m_file.size == m_sentBytes; }
+  bool ackFinished() const { return m_seq > 0 && m_seq == m_ack_seq; }
 
 private:
   JID m_friendId;
@@ -55,21 +55,24 @@ private:
   const IM *m_im;
   int m_buf = 4096; // 4k
   int m_seq = 0;
-  int m_sentBytes = 0;
+  quint64 m_sentBytes = 0;
   int m_ack_seq = 0;
 
   std::unique_ptr<InBandBytestream> m_ibb;
   std::unique_ptr<QFile> qFile;
   gloox::Bytestream *m_byteStream;
+
 public:
 signals:
   void fileSent(const JID &m_friendId, const FileHandler::File &m_file);
-  void fileError(const JID &m_friendId, const FileHandler::File &m_file, int m_sentBytes);
-  void fileAbort(const JID &m_friendId, const FileHandler::File &m_file, int m_sentBytes);
+  void fileError(const JID &m_friendId, const FileHandler::File &m_file,
+                 int m_sentBytes);
+  void fileAbort(const JID &m_friendId, const FileHandler::File &m_file,
+                 int m_sentBytes);
   void fileSending(const JID &m_friendId, const FileHandler::File &m_file,
                    int m_seq, int m_sentBytes, bool end);
 };
 
-} // namespace IM
+} // namespace messenger
 } // namespace lib
 #endif // OKEDU_CLASSROOM_DESKTOP_IMFILE_H
